Moved Status.toCsv due to circular dependencies.
