+++ Order/StatusList.ts
@@ -12,4 +12,12 @@ export namespace StatusList {
 			)
 		)
 	}
+	export function toCsv(value: Status[] | StatusList | undefined): string {
+		let result = ``
+		if (Array.isArray(value))
+			result += `"` + value.join(" ") + `"`
+		else if (StatusList.is(value))
+			result = Object.entries(value).reduce((l, p) => (l += p), "")
+		return result
+	}
 }
